+++ src/Libraries/SmartStore.Data/Migrations/201809171309522_NewsletterSubscriptionLanguage.cs
@@ -1,0 +1,18 @@
+namespace SmartStore.Data.Migrations
+{
+    using System;
+    using System.Data.Entity.Migrations;
+    
+    public partial class NewsletterSubscriptionLanguage : DbMigration
+    {
+        public override void Up()
+        {
+            AddColumn("dbo.NewsLetterSubscription", "WorkingLanguageId", c => c.Int(nullable: false));
+        }
+        
+        public override void Down()
+        {
+            DropColumn("dbo.NewsLetterSubscription", "WorkingLanguageId");
+        }
+    }
+}