+++ src/Libraries/SmartStore.Services/Search/Forum/ForumSearchResult.cs
@@ -9,14 +9,14 @@ namespace SmartStore.Services.Search
 {
     public partial class ForumSearchResult
     {
-        private readonly Func<IList<ForumTopic>> _hitsFactory;
-		private IPagedList<ForumTopic> _hits;
+        private readonly Func<IList<ForumPost>> _hitsFactory;
+		private IPagedList<ForumPost> _hits;
 
 		public ForumSearchResult(
 			ISearchEngine engine,
             ForumSearchQuery query,
 			int totalHitsCount,
-			Func<IList<ForumTopic>> hitsFactory,
+			Func<IList<ForumPost>> hitsFactory,
 			string[] spellCheckerSuggestions,
             IDictionary<string, FacetGroup> facets)
 		{
@@ -27,7 +27,7 @@ namespace SmartStore.Services.Search
 			SpellCheckerSuggestions = spellCheckerSuggestions ?? new string[0];
             Facets = facets ?? new Dictionary<string, FacetGroup>();
 
-            _hitsFactory = hitsFactory ?? (() => new List<ForumTopic>());
+            _hitsFactory = hitsFactory ?? (() => new List<ForumPost>());
 			TotalHitsCount = totalHitsCount;
 		}
 
@@ -36,24 +36,24 @@ namespace SmartStore.Services.Search
         /// </summary>
         /// <param name="query">Forum search query</param>
         public ForumSearchResult(ForumSearchQuery query)
-			: this(null, query, 0, () => new List<ForumTopic>(), null, null)
+			: this(null, query, 0, () => new List<ForumPost>(), null, null)
 		{
 		}
 
         /// <summary>
-        /// Forum topics found
+        /// Forum posts found.
         /// </summary>
-        public IPagedList<ForumTopic> Hits
+        public IPagedList<ForumPost> Hits
 		{
 			get
 			{
 				if (_hits == null)
 				{
 					var entities = TotalHitsCount == 0 
-						? new List<ForumTopic>() 
+						? new List<ForumPost>() 
 						: _hitsFactory.Invoke();
 
-					_hits = new PagedList<ForumTopic>(entities, Query.PageIndex, Query.Take, TotalHitsCount);
+					_hits = new PagedList<ForumPost>(entities, Query.PageIndex, Query.Take, TotalHitsCount);
 				}
 
 				return _hits;
@@ -63,12 +63,12 @@ namespace SmartStore.Services.Search
         public int TotalHitsCount { get; }
 
         /// <summary>
-        /// The original forum search query
+        /// The original forum search query.
         /// </summary>
         public ForumSearchQuery Query { get; private set; }
 
 		/// <summary>
-		/// Gets spell checking suggestions/corrections
+		/// Gets spell checking suggestions/corrections.
 		/// </summary>
 		public string[] SpellCheckerSuggestions { get; set;	}