+++ src/Presentation/SmartStore.Web/Models/Common/CommentModel.cs
@@ -1,14 +1,12 @@
 ﻿using System;
-using System.Linq;
 using System.Collections.Generic;
 using SmartStore.Web.Framework.Modelling;
+using SmartStore.Web.Models.Customer;
 
 namespace SmartStore.Web.Models.Common
 {
-	public partial class CommentListModel : ModelBase
+    public partial class CommentListModel : ModelBase
 	{
-		private bool? _hasAnyAvatar;
-
 		public CommentListModel()
 		{
 			Comments = new List<CommentModel>();
@@ -18,17 +16,6 @@ namespace SmartStore.Web.Models.Common
 		public int NumberOfComments { get; set; }
 		public IList<CommentModel> Comments { get; set; }
 		public bool AllowCustomersToUploadAvatars { get; set; }
-		public int AvatarPictureSize { get; set; }
-
-		public bool HasAnyAvatar()
-		{
-			if (_hasAnyAvatar == null)
-			{
-				_hasAnyAvatar = this.AllowCustomersToUploadAvatars && Comments.Any(x => x.CustomerAvatarUrl.HasValue());
-			}
-
-			return _hasAnyAvatar.Value;
-		}
 	}
 
 	public partial class CommentModel : EntityModelBase
@@ -39,24 +26,24 @@ namespace SmartStore.Web.Models.Common
 		{
 			Guard.NotNull(parent, nameof(parent));
 
+            Avatar = new CustomerAvatarModel();
 			_parent = new WeakReference<CommentListModel>(parent);
 		}
 
 		public int CustomerId { get; set; }
         public string CustomerName { get; set; }
-        public string CustomerAvatarUrl { get; set; }
         public string CommentTitle { get; set; }
         public string CommentText { get; set; }
         public DateTime CreatedOn { get; set; }
 		public string CreatedOnPretty { get; set; }
 		public bool AllowViewingProfiles { get; set; }
+        public CustomerAvatarModel Avatar { get; set; }
 
-		public CommentListModel Parent
+        public CommentListModel Parent
 		{
 			get
 			{
-				CommentListModel parent;
-				_parent.TryGetTarget(out parent);
+				_parent.TryGetTarget(out var parent);
 				return parent;
 			}
 		}