+++ src/Presentation/SmartStore.Web/Models/Profile/ProfileIndexModel.cs
@@ -1,9 +1,9 @@
-﻿
+﻿using SmartStore.Web.Framework.Modelling;
+
 namespace SmartStore.Web.Models.Profile
 {
-    public partial class ProfileIndexModel
+    public partial class ProfileIndexModel : EntityModelBase
     {
-        public int CustomerProfileId { get; set; }
         public string ProfileTitle { get; set; }
         public int PostsPage { get; set; }
         public bool PagingPosts { get; set; }